behemoth fix

+++ data/behemoth.h
@@ -32,8 +32,8 @@ typedef signed char INT8;
 typedef void VOID;
 typedef char FILLER;
 //This may be 64 bit in later phases, but in the PEI phase it is confirmed to be 32 bits
-//typedef UINT64 UINTN;
-typedef UINT32 UINTN;
+typedef UINT64 UINTN;
+//typedef UINT32 UINTN;
 
 typedef INT64 INTN;
